+++ lexicmap/cmd/genomes.go
@@ -1,0 +1,96 @@
+// Copyright © 2023-2024 Wei Shen <shenwei356@gmail.com>
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+package cmd
+
+import (
+	"fmt"
+	"path/filepath"
+	"strings"
+
+	"github.com/shenwei356/bio/seq"
+	"github.com/spf13/cobra"
+	"github.com/twotwotwo/sorts/sortutil"
+)
+
+var genomesCmd = &cobra.Command{
+	Use:   "genomes",
+	Short: "View genome IDs in the index",
+	Long: `View genome IDs in the index
+
+`,
+	Run: func(cmd *cobra.Command, args []string) {
+		opt := getOptions(cmd)
+		seq.ValidateSeq = false
+
+		// ------------------------------
+
+		dbDir := getFlagString(cmd, "index")
+		if dbDir == "" {
+			checkError(fmt.Errorf("flag -d/--index needed"))
+		}
+
+		outFile := getFlagString(cmd, "out-file")
+
+		// ---------------------------------------------------------------
+
+		// genomes.map file for mapping index to genome id
+		m, err := readGenomeMapName2Idx(filepath.Join(dbDir, FileGenomeIndex))
+		if err != nil {
+			checkError(fmt.Errorf("failed to read genomes index mapping file: %s", err))
+		}
+
+		// output file handler
+		outfh, gw, w, err := outStream(outFile, strings.HasSuffix(outFile, ".gz"), opt.CompressionLevel)
+		checkError(err)
+		defer func() {
+			outfh.Flush()
+			if gw != nil {
+				gw.Close()
+			}
+			w.Close()
+		}()
+
+		ids := make([]string, len(m))
+		i := 0
+		for id := range m {
+			ids[i] = id
+			i++
+		}
+		sortutil.Strings(ids)
+
+		for _, id := range ids {
+			outfh.WriteString(id)
+			outfh.WriteString("\n")
+		}
+	},
+}
+
+func init() {
+	utilsCmd.AddCommand(genomesCmd)
+
+	genomesCmd.Flags().StringP("index", "d", "",
+		formatFlagUsage(`Index directory created by "lexicmap index".`))
+
+	genomesCmd.Flags().StringP("out-file", "o", "-",
+		formatFlagUsage(`Out file, supports the ".gz" suffix ("-" for stdout).`))
+
+	genomesCmd.SetUsageTemplate(usageTemplate(""))
+}