+++ CHANGELOG.md
@@ -24,6 +24,7 @@
     - `lexicmap utils gen-masks` for generating masks from the top N largest genomes.
     - `lexicmap utils seed-pos` for extracting seed positions via reference names.
     - `lexicmap utils reindex-seeds` for recreating indexes of k-mer-value (seeds) data.
+    - `lexicmap utils genomes` for list genomes IDs in the index.
 
 ### v0.2.0 - 2024-02-02